Refactor: 디자인 패턴 단위 페이지 별 폴더 추가 컴포넌트가 많아질 거 같아서 molecules과 organism에 페이지 별 폴더를 추가하여 컴포넌트를 넣었으며 공통으로 사용할 컴포넌트는 폴더 밖으로 뺐다. 더미리스트도 삭제하였다.
