+++ src/components/template/ChatRoomTemplate.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import ChatHeader from "../organism/ChatHeader";
-import ChatArea from "../organism/ChatArea";
-import ChatMessageInput from "../organism/ChatMessageInput";
+import ChatHeader from "../organism/chatroom/ChatHeader";
+import ChatArea from "../organism/chatroom/ChatArea";
+import ChatMessageInput from "../organism/chatroom/ChatMessageInput";
 
 function ChatRoomTemplate() {
   return (
